城市列表相关操作弄到仓库中
