+++ src/views/City/index.vue
@@ -50,8 +50,7 @@
 
 <script>
 import NormalHeader from '../../components/NormalHeader'
-import { mapMutations, mapGetters } from 'vuex'
-import { getCityList } from '@/api/city'
+import { mapMutations, mapGetters, mapActions } from 'vuex'
 
 export default {
   name: 'City',
@@ -60,66 +59,13 @@ export default {
     NormalHeader
   },
 
-  data () {
-    return {
-      cities: []
-    }
-  },
-
   computed: {
-    ...mapGetters('city', ['curCityName']),
-
-    cityList () {
-      // 最终结果收集
-      const result = []
-      this.cities.forEach(item => {
-        // 获取当前城市的首字母
-        const py = item.pinyin.substr(0, 1).toUpperCase()
-
-        // 判断 当前 py 是否已经在 result 中存在，如果存在这个 py 在 result 中的下标是多少呢？
-        const index = result.findIndex(item => item.py === py)
-        if (index > -1) {
-          // 存在
-          result[index].list.push(item)
-        } else {
-          // 不存在
-          result.push({
-            py,
-            list: [item]
-          })
-        }
-        // 1. [{ py: 'B', list: [北京-item] }]
-        // 2. 保定 [{ py: 'B', list: [北京item] }, { py: 'B', list: [保定-item]}]
-        // 想要 [{ py: 'B', list: [北京-item, 保定-item] }]
-      })
-      return result.sort((a, b) => {
-        return a.py.charCodeAt() - b.py.charCodeAt()
-      })
-    },
-
-    indexs () {
-      return this.cityList.map(item => item.py)
-    }
+    ...mapGetters('city', ['curCityName', 'cityList', 'indexs'])
   },
 
   methods: {
+    ...mapActions('city', ['GET_CITIES']),
     ...mapMutations('city', ['SET_CURCITY']),
-
-    getCityList () {
-      getCityList()
-        .then(res => {
-          const data = res.data
-          if (data.status === 0) {
-            this.cities = data.data.cities
-          } else {
-            alert(data.msg)
-          }
-        })
-        .catch(err => {
-          console.log(err)
-          alert('网络异常，请稍后重试')
-        })
-    },
 
     handleIndex (py) {
       // ref 标记时，如果是标记在 v-for 上。那么得到的是一个数组
@@ -141,7 +87,7 @@ export default {
   },
 
   created () {
-    this.getCityList()
+    this.GET_CITIES()
   }
 }
 </script>
